# backend/swagger.yaml
openapi: 3.0.3
info:
  title: Performance-Based Salary Generator API
  description: |
    A comprehensive REST API for managing performance-based salary calculations.
    
    ## Features
    - **Authentication**: JWT-based authentication with role-based access control
    - **User Management**: Admin and employee roles with different permissions
    - **Performance Tracking**: Record and manage employee performance metrics
    - **Salary Calculation**: Automated salary calculation based on performance
    - **Salary History**: Track and view salary history and breakdowns
    - **Settings Management**: Global application settings and statistics
    
    ## Authentication
    Most endpoints require authentication. Include the JWT token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Roles
    - **Admin**: Full access to all endpoints and data
    - **Employee**: Limited access to own data and performance records
    
    ## Base URL
    ```
    http://localhost:5000/api
    ```
  version: 1.0.0
  contact:
    name: API Support
    email: support@company.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000/api
    description: Development server
  - url: https://api.company.com/api
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
  schemas:
    User:
      type: object
      properties:
        _id:
          type: string
          description: Unique user identifier
        name:
          type: string
          description: User's full name
        email:
          type: string
          format: email
          description: User's email address
        role:
          type: string
          enum: [admin, employee]
          description: User's role in the system
        baseSalary:
          type: number
          description: User's base salary
        department:
          type: string
          description: User's department
        isActive:
          type: boolean
          description: Whether the user account is active
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - name
        - email
        - role
        - baseSalary
        - department
    Performance:
      type: object
      properties:
        _id:
          type: string
          description: Unique performance record identifier
        employeeId:
          $ref: '#/components/schemas/User'
        taskName:
          type: string
          description: Name of the task performed
        score:
          type: number
          minimum: 0
          maximum: 100
          description: Performance score (0-100)
        difficulty:
          type: string
          enum: [easy, medium, hard]
          description: Task difficulty level
        clientFeedback:
          type: string
          description: Feedback from client
        date:
          type: string
          format: date-time
          description: Date of performance
        notes:
          type: string
          description: Additional notes
        isApproved:
          type: boolean
          description: Whether the record is approved
        difficultyMultiplier:
          type: number
          description: Calculated difficulty multiplier
        weightedScore:
          type: number
          description: Calculated weighted score
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - employeeId
        - taskName
        - score
        - difficulty
...
    Settings:
      type: object
      properties:
        _id:
          type: string
          description: Unique settings identifier
        salaryMultiplier:
          type: number
          description: Global salary multiplier
        lastSalaryCalculation:
          type: string
          format: date-time
          nullable: true
        systemStats:
          type: object
          properties:
            totalEmployees:
              type: number
            totalPerformanceRecords:
              type: number
            totalSalariesGenerated:
              type: number
            lastUpdated:
              type: string
              format: date-time
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Error:
      type: object
      properties:
        message:
          type: string
          description: Error message
        error:
          type: string
          description: Error type
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
              value:
                type: string
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        data:
          type: object
    Pagination:
      type: object
      properties:
        page:
          type: integer
          description: Current page number
        limit:
          type: integer
          description: Number of items per page
        total:
          type: integer
          description: Total number of items
        pages:
          type: integer
          description: Total number of pages
paths:
...
  /users:
    get:
      summary: Get all users (Admin only)
      description: Retrieve a paginated list of all users
      tags:
        - Users
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: Page number
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        '200':
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          users:
                            type: array
                            items:
                              $ref: '#/components/schemas/User'
                          pagination:
                            $ref: '#/components/schemas/Pagination'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /performance:
    get:
      summary: Get performance records
      description: Retrieve performance records (admin sees all, employee sees own)
      tags:
        - Performance
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: Page number
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: startDate
          in: query
          description: Start date filter (ISO 8601)
          required: false
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          description: End date filter (ISO 8601)
          required: false
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Performance records retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          records:
                            type: array
                            items:
                              $ref: '#/components/schemas/Performance'
                          pagination:
                            $ref: '#/components/schemas/Pagination'

    post:
      summary: Create performance record
      description: Create a new performance record
      tags:
        - Performance
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - taskName
                - score
                - difficulty
              properties:
                employeeId:
                  type: string
                  description: Employee ID (admin only, defaults to current user)
                taskName:
                  type: string
                  minLength: 2
                  maxLength: 200
                  example: Frontend Development - User Dashboard
                score:
                  type: number
                  minimum: 0
                  maximum: 100
                  example: 85
                difficulty:
                  type: string
                  enum: [easy, medium, hard]
                  example: medium
                clientFeedback:
                  type: string
                  maxLength: 500
                  example: Great work on the dashboard design
                date:
                  type: string
                  format: date-time
                  example: 2025-01-15T10:30:00Z
                notes:
                  type: string
                  maxLength: 1000
                  example: Completed ahead of schedule
      responses:
        '201':
          description: Performance record created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          record:
                            $ref: '#/components/schemas/Performance'
...
  /settings:
    get:
      summary: Get current settings (Admin only)
      description: Retrieve current application settings
      tags:
        - Settings
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Settings retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          settings:
                            $ref: '#/components/schemas/Settings'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: Update settings (Admin only)
      description: Update application settings
      tags:
        - Settings
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                salaryMultiplier:
                  type: number
                  minimum: 0
                  example: 100
      responses:
        '200':
          description: Settings updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          settings:
                            $ref: '#/components/schemas/Settings'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...